fix(properties): GeneratorProperties
Use the same properties for the uid workspace between the
WorksapceUidProperties and the GeneratorProperties

This was causing issues and the workspace uid path was not honored and
tmp files were created in several places between a default
/tmp/workspace-uid and the one defined in WorksapceUidProperties.

I also removed the configuration of the extractPath inside the uid
workspace as it seems not relevant to specialize this path.

livBuildEnabled was also not set on the GeneratorProperties.

I've made i18n resource path static also as I did not see the use case
to have it as a property of GeneratorProperties.

I feel that WorksapceUidProperties should be shared between artifact
builder module and angularjs generator to avoid having big constructors.